+++ src/apify_client/clients/resource_clients/request_queue.py
@@ -1,7 +1,7 @@
 from typing import Any, Dict, Optional
 
 from ..._errors import ApifyApiError
-from ..._utils import _catch_not_found_or_throw, _parse_date_fields, _pluck_data
+from ..._utils import _catch_not_found_or_throw, _filter_out_none_values_recursively, _parse_date_fields, _pluck_data
 from ..base import ResourceClient
 
 
@@ -39,11 +39,11 @@ class RequestQueueClient(ResourceClient):
         Returns:
             dict: The updated request queue
         """
-        updated_fields = {}
-        if name is not None:
-            updated_fields['name'] = name
+        updated_fields = {
+            'name': name,
+        }
 
-        return self._update(updated_fields)
+        return self._update(_filter_out_none_values_recursively(updated_fields))
 
     def delete(self) -> None:
         """Delete the request queue.