+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Changelog
 - simplified development scripts
 - updated GitHub Actions versions to fix deprecations
 - unified unit test style
+- unified preparing resource representation
 
 [0.6.0](../../releases/tag/v0.6.0) - 2022-06-27
 -----------------------------------------------